drivers: refactor: move interrupt from platform to drivers
