+++ src/drivers/intel/baytrail/interrupt.c
