+++ src/drivers/intel/cavs/Makefile.am
@@ -4,6 +4,7 @@ libdrivers_la_SOURCES = \
 	dmic.c \
 	hda-dma.c \
 	hda.c \
+	interrupt.c \
 	ssp.c
 
 libdrivers_la_CFLAGS = \
