+++ src/include/sof/drivers/Makefile.am
@@ -1,0 +1,2 @@
+noinst_HEADERS = \
+	interrupt.h
