+++ src/platform/baytrail/Makefile.am
@@ -5,7 +5,6 @@ EXTRA_DIST = baytrail.x.in
 noinst_LTLIBRARIES = libplatform.la
 
 libplatform_la_SOURCES = \
-	interrupt.c \
 	platform.c \
 	dai.c \
 	dma.c \
