+++ src/platform/baytrail/include/platform/interrupt.h
@@ -88,13 +88,4 @@
 /* no nested interrupts */
 #define PLATFORM_IRQ_CHILDREN	0
 
-void platform_interrupt_init(void);
-
-struct irq_desc *platform_irq_get_parent(uint32_t irq);
-void platform_interrupt_set(int irq);
-void platform_interrupt_clear(uint32_t irq, uint32_t mask);
-uint32_t platform_interrupt_get_enabled(void);
-void platform_interrupt_mask(uint32_t irq, uint32_t mask);
-void platform_interrupt_unmask(uint32_t irq, uint32_t mask);
-
 #endif
