+++ src/platform/haswell/include/platform/interrupt.h
@@ -66,13 +66,4 @@
 #define IRQ_MASK_EXT_PARITY	(1 << IRQ_NUM_EXT_PARITY)
 #define IRQ_MASK_TIMER3		(1 << IRQ_NUM_TIMER3)
 
-void platform_interrupt_init(void);
-
-struct irq_desc *platform_irq_get_parent(uint32_t irq);
-void platform_interrupt_set(int irq);
-void platform_interrupt_clear(uint32_t irq, uint32_t mask);
-uint32_t platform_interrupt_get_enabled(void);
-void platform_interrupt_mask(uint32_t irq, uint32_t mask);
-void platform_interrupt_unmask(uint32_t irq, uint32_t mask);
-
 #endif
